Refactor : 서버와 통신하는 로직 분리
